+++ __init__.py
@@ -1,12 +1,18 @@
-from scrappers.engine.config import Configuration
-from scrappers.engine.user_agent import (
-    get_rand_agent, get_rand_unique_agent, get_user_agent, get_user_agents)
-from scrappers.images.picturepub import PicturePub
-from scrappers.images.sawfirst import SawFirst
-from scrappers.tennis.matchstats.parser import MatchStats
-from scrappers.tennis.wta.parser import ParsePage
-from scrappers.volleyball.base_volleyball import (PlayerPage, TeamPage,
-                                                  TeamsPage)
-from scrappers.engine.utilities import guess_celebrity
-
-config = Configuration()
+from scrappers.scrappers.config import utilities
+from scrappers.scrappers.config.config import Configuration, configuration
+from scrappers.scrappers.config.http import user_agent
+from scrappers.scrappers.config.http.aws import (
+    QueryManager, TransferManager, create_object_url, image_size_creator,
+    unique_path_creator)
+from scrappers.scrappers.config.http.engine import Requestor
+from scrappers.scrappers.config.http.user_agent import (AGENTS_LIST,
+                                                        get_rand_agent,
+                                                        get_rand_unique_agent,
+                                                        get_user_agent,
+                                                        get_user_agents)
+from scrappers.scrappers.tennis.matchstats.parser import MatchStats
+from scrappers.scrappers.tennis.wta.models import (Player, Tournament,
+                                                   TournamentMatch)
+from scrappers.scrappers.tennis.wta.parser import ParsePage
+from scrappers.scrappers.volleyball.parser import (PlayerPage, TeamPage,
+                                                   TeamsPage)