+++ scrappers/volleyball/models.py
@@ -1,0 +1,8 @@
+from collections import namedtuple
+
+# ENHANCEMENT: Subclass a list instead?
+class Player(namedtuple('Player', ['name', 'link', 'date_of_birth',
+                         'age', 'height', 'weight', 'spike', 'block'])):
+    """Player class used to create a volleyball player
+    """
+    __slots__ = ()